+++ packages/cli/snap-tests-global/new-create-vite-migrates-eslint-prettier/snap.txt
@@ -1,0 +1,135 @@
+> vp create vite@9.0.5 --no-interactive -- my-react-ts --template react-ts # create vite app with pinned version + react-ts template, should auto-migrate ESLint -> Oxlint and merge lint config into vite.config.ts
+> test ! -f my-react-ts/eslint.config.js && echo 'eslint.config.js removed' # eslint config deleted
+eslint.config.js removed
+
+> test ! -f my-react-ts/.oxlintrc.json && echo '.oxlintrc.json merged into vite.config.ts' # migration output merged by rewrite step (matches vp migrate)
+.oxlintrc.json merged into vite.config.ts
+
+> cat my-react-ts/vite.config.ts # merged vite config should contain lint and fmt sections
+import { defineConfig } from "vite-plus";
+import react from "@vitejs/plugin-react";
+
+// https://vite.dev/config/
+export default defineConfig({
+  staged: {
+    "*": "vp check --fix",
+  },
+  fmt: {},
+  lint: {
+    plugins: ["oxc", "typescript", "unicorn", "react"],
+    categories: {
+      correctness: "warn",
+    },
+    env: {
+      builtin: true,
+    },
+    ignorePatterns: ["dist"],
+    overrides: [
+      {
+        files: ["**/*.{ts,tsx}"],
+        rules: {
+          "constructor-super": "error",
+          "for-direction": "error",
+          "getter-return": "error",
+          "no-async-promise-executor": "error",
+          "no-case-declarations": "error",
+          "no-class-assign": "error",
+          "no-compare-neg-zero": "error",
+          "no-cond-assign": "error",
+          "no-const-assign": "error",
+          "no-constant-binary-expression": "error",
+          "no-constant-condition": "error",
+          "no-control-regex": "error",
+          "no-debugger": "error",
+          "no-delete-var": "error",
+          "no-dupe-class-members": "error",
+          "no-dupe-else-if": "error",
+          "no-dupe-keys": "error",
+          "no-duplicate-case": "error",
+          "no-empty": "error",
+          "no-empty-character-class": "error",
+          "no-empty-pattern": "error",
+          "no-empty-static-block": "error",
+          "no-ex-assign": "error",
+          "no-extra-boolean-cast": "error",
+          "no-fallthrough": "error",
+          "no-func-assign": "error",
+          "no-global-assign": "error",
+          "no-import-assign": "error",
+          "no-invalid-regexp": "error",
+          "no-irregular-whitespace": "error",
+          "no-loss-of-precision": "error",
+          "no-misleading-character-class": "error",
+          "no-new-native-nonconstructor": "error",
+          "no-nonoctal-decimal-escape": "error",
+          "no-obj-calls": "error",
+          "no-prototype-builtins": "error",
+          "no-redeclare": "error",
+          "no-regex-spaces": "error",
+          "no-self-assign": "error",
+          "no-setter-return": "error",
+          "no-shadow-restricted-names": "error",
+          "no-sparse-arrays": "error",
+          "no-this-before-super": "error",
+          "no-undef": "error",
+          "no-unexpected-multiline": "error",
+          "no-unreachable": "error",
+          "no-unsafe-finally": "error",
+          "no-unsafe-negation": "error",
+          "no-unsafe-optional-chaining": "error",
+          "no-unused-labels": "error",
+          "no-unused-private-class-members": "error",
+          "no-unused-vars": "error",
+          "no-useless-backreference": "error",
+          "no-useless-catch": "error",
+          "no-useless-escape": "error",
+          "no-with": "error",
+          "require-yield": "error",
+          "use-isnan": "error",
+          "valid-typeof": "error",
+          "no-array-constructor": "error",
+          "no-unused-expressions": "error",
+          "typescript/ban-ts-comment": "error",
+          "typescript/no-duplicate-enum-values": "error",
+          "typescript/no-empty-object-type": "error",
+          "typescript/no-explicit-any": "error",
+          "typescript/no-extra-non-null-assertion": "error",
+          "typescript/no-misused-new": "error",
+          "typescript/no-namespace": "error",
+          "typescript/no-non-null-asserted-optional-chain": "error",
+          "typescript/no-require-imports": "error",
+          "typescript/no-this-alias": "error",
+          "typescript/no-unnecessary-type-constraint": "error",
+          "typescript/no-unsafe-declaration-merging": "error",
+          "typescript/no-unsafe-function-type": "error",
+          "typescript/no-wrapper-object-types": "error",
+          "typescript/prefer-as-const": "error",
+          "typescript/prefer-namespace-keyword": "error",
+          "typescript/triple-slash-reference": "error",
+          "react/rules-of-hooks": "error",
+          "react/exhaustive-deps": "warn",
+          "react/only-export-components": [
+            "error",
+            {
+              allowConstantExport: true,
+            },
+          ],
+        },
+        env: {
+          es2020: true,
+          browser: true,
+        },
+      },
+    ],
+    options: {
+      typeAware: true,
+      typeCheck: true,
+    },
+  },
+  plugins: [react()],
+});
+
+> node -e "const p=require('./my-react-ts/package.json');console.log('lint:', p.scripts && p.scripts.lint);console.log('eslint dep:', !!(p.devDependencies && p.devDependencies.eslint));console.log('vite-plus dep:', !!(p.devDependencies && p.devDependencies['vite-plus']));" # scripts rewritten, eslint dep removed, vite-plus added
+lint: vp lint .
+eslint dep: false
+vite-plus dep: true
