+++ packages/cli/src/migration/bin.ts
@@ -43,6 +43,8 @@ import {
   addFrameworkShim,
   checkVitestVersion,
   checkViteVersion,
+  confirmEslintMigration,
+  confirmPrettierMigration,
   detectEslintProject,
   detectFramework,
   detectNodeVersionManagerFile,
@@ -54,135 +56,17 @@ import {
   migrateNodeVersionManagerFile,
   migratePrettierToOxfmt,
   preflightGitHooksSetup,
+  promptEslintMigration,
+  promptPrettierMigration,
   rewriteMonorepo,
   rewriteStandaloneProject,
+  warnLegacyEslintConfig,
+  warnPackageLevelEslint,
+  warnPackageLevelPrettier,
   type Framework,
   type NodeVersionManagerDetection,
 } from './migrator.ts';
 import { createMigrationReport, type MigrationReport } from './report.ts';
-
-function warnPackageLevelEslint() {
-  prompts.log.warn(
-    'ESLint detected in workspace packages but no root config found. Package-level ESLint must be migrated manually.',
-  );
-}
-
-function warnLegacyEslintConfig(legacyConfigFile: string) {
-  prompts.log.warn(
-    `Legacy ESLint configuration detected (${legacyConfigFile}). ` +
-      'Automatic migration to Oxlint requires ESLint v9+ with flat config format (eslint.config.*). ' +
-      'Please upgrade to ESLint v9 first: https://eslint.org/docs/latest/use/migrate-to-9.0.0',
-  );
-}
-
-async function confirmEslintMigration(interactive: boolean): Promise<boolean> {
-  if (interactive) {
-    const confirmed = await prompts.confirm({
-      message:
-        'Migrate ESLint rules to Oxlint using @oxlint/migrate?\n  ' +
-        styleText(
-          'gray',
-          "Oxlint is Vite+'s built-in linter — significantly faster than ESLint with compatible rule support. @oxlint/migrate converts your existing rules automatically.",
-        ),
-      initialValue: true,
-    });
-    if (prompts.isCancel(confirmed)) {
-      cancelAndExit();
-    }
-    return confirmed;
-  }
-  return true;
-}
-
-async function promptEslintMigration(
-  projectPath: string,
-  interactive: boolean,
-  packages?: WorkspacePackage[],
-): Promise<boolean> {
-  const eslintProject = detectEslintProject(projectPath, packages);
-  if (eslintProject.hasDependency && !eslintProject.configFile && eslintProject.legacyConfigFile) {
-    warnLegacyEslintConfig(eslintProject.legacyConfigFile);
-    return false;
-  }
-  if (!eslintProject.hasDependency) {
-    return false;
-  }
-  if (!eslintProject.configFile) {
-    // Packages have eslint but no root config → warn and skip
-    warnPackageLevelEslint();
-    return false;
-  }
-  const confirmed = await confirmEslintMigration(interactive);
-  if (!confirmed) {
-    return false;
-  }
-  const ok = await migrateEslintToOxlint(
-    projectPath,
-    interactive,
-    eslintProject.configFile,
-    packages,
-  );
-  if (!ok) {
-    cancelAndExit('ESLint migration failed. Fix the issue and re-run `vp migrate`.', 1);
-  }
-  return true;
-}
-
-function warnPackageLevelPrettier() {
-  prompts.log.warn(
-    'Prettier detected in workspace packages but no root config found. Package-level Prettier must be migrated manually.',
-  );
-}
-
-async function confirmPrettierMigration(interactive: boolean): Promise<boolean> {
-  if (interactive) {
-    const confirmed = await prompts.confirm({
-      message:
-        'Migrate Prettier to Oxfmt?\n  ' +
-        styleText(
-          'gray',
-          "Oxfmt is Vite+'s built-in formatter that replaces Prettier with faster performance. Your configuration will be converted automatically.",
-        ),
-      initialValue: true,
-    });
-    if (prompts.isCancel(confirmed)) {
-      cancelAndExit();
-    }
-    return confirmed;
-  }
-  prompts.log.info('Prettier configuration detected. Auto-migrating to Oxfmt...');
-  return true;
-}
-
-async function promptPrettierMigration(
-  projectPath: string,
-  interactive: boolean,
-  packages?: WorkspacePackage[],
-): Promise<boolean> {
-  const prettierProject = detectPrettierProject(projectPath, packages);
-  if (!prettierProject.hasDependency) {
-    return false;
-  }
-  if (!prettierProject.configFile) {
-    // Packages have prettier but no root config → warn and skip
-    warnPackageLevelPrettier();
-    return false;
-  }
-  const confirmed = await confirmPrettierMigration(interactive);
-  if (!confirmed) {
-    return false;
-  }
-  const ok = await migratePrettierToOxfmt(
-    projectPath,
-    interactive,
-    prettierProject.configFile,
-    packages,
-  );
-  if (!ok) {
-    cancelAndExit('Prettier migration failed. Fix the issue and re-run `vp migrate`.', 1);
-  }
-  return true;
-}
 
 async function confirmNodeVersionFileMigration(
   interactive: boolean,
